fixed safe to import as a module
watch out for `__` on globals, when importing, that private scope is not visible to functions within the module, not a bug when running `python mymod.py`, but when `import mymod` this issue pops up. PYTHON GOTCHA, so basically never use `__` prefix for globals unless you know what you are doing.